Added support for try-with-resources
Since Java 7 there is an advanced syntax to automatically close
`AutoCloseable` types, namely `try-with-resources statement`. This
commit adds the ability to convert functions which consume an
`AutoClosable` into functions which automatically close the
`AutoCloseable` after exuction. Further, the return type is wrapped in a
`Try`. Hence, this commit enables the `Try` type to act like the
`try-with-resources statment` and automatically close `AutoCloseables`.